feat: add pending notification, store query params
